src/backend/FastAPI/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
psycopg2-binary==2.9.10
